Add Criterion benchmarks
Introduces a (hopefully) representative set of benchmarks for an
imaginary 'User session token', with a variable size depending on which
fields are populated. Does not yet cover floating point or union types.

Also included is a go-bare implementation of an identical benchmark, so
that we can compare our implementation again the 'reference' code.